requirements.txt
oandapyV20>=0.7.2
httpx>=0.27
aiohttp>=3.9
pydantic>=2.5
numpy>=1.26
pandas>=2.1
scikit-learn>=1.4
lightgbm>=4.2
python-dotenv>=1.0
loguru>=0.7
rich>=13.7
tenacity>=8.2
ta>=0.11.0
